Aligning the Category boxes with the grid.
The left and right margins of the Category boxes have now been adjusted to
align the boxes with the rest of the grid. To do this, specific groups were
added for iotaleft, iotacenter and iotaright. It may have been more elegant
to adjust the properties of the parent container, but this was a more local
change.